refactor: move test directory to root from src/
